+++ app/views/providers/home_address/home_address_lookups/show.html.erb
@@ -1,7 +1,7 @@
 <%= form_with(
       model: @form,
       scope: :address_lookup,
-      url: providers_legal_aid_application_home_address_lookup_path(@legal_aid_application),
+      url: providers_legal_aid_application_home_address_home_address_lookup_path(@legal_aid_application),
       method: :patch,
       local: true,
     ) do |form| %>
