--- app/views/providers/home_address_selections/show.html.erb
@@ -1,16 +1,0 @@
-<%= form_with(
-      model: @form,
-      scope: :address_selection,
-      url: providers_legal_aid_application_home_address_selection_path(@legal_aid_application),
-      method: :patch,
-      local: true,
-    ) do |form| %>
-
-    <% if @address_collection.count > 1 %>
-      <%= render "shared/home_address_selection/select_address", form:, addresses: @addresses %>
-    <% elsif @address_collection.count == 1 %>
-      <%= render "shared/home_address_selection/confirm_address", form: %>
-    <% else %>
-      <%= render "shared/home_address_selection/no_address", form: %>
-    <% end %>
-<% end %>
