+++ app/views/shared/address_selection/_select_address.html.erb
@@ -1,5 +1,5 @@
 <%= page_template(
-      page_title: t(".heading"),
+      page_title: t(".heading", location:),
       template: :basic,
       form:,
     ) do %>
