--- app/views/shared/home_address_selection/_confirm_address.html.erb
@@ -1,47 +1,0 @@
-<%= page_template(
-      page_title: t(".heading"),
-      template: :basic,
-      form:,
-    ) do %>
-
-      <h1 class="govuk-heading-xl"><%= t(".heading") %></h1>
-
-      <p class='govuk-body'>
-        <% if @form.model.building_number_name.present? %>
-          <%= t(".addresses_found_long_html",
-                number: @address_collection.count,
-                postcode: @form.model.pretty_postcode,
-                building_name_number: @form.model.building_number_name) %>
-        <% else %>
-          <%= t(".addresses_found_html",
-                number: @address_collection.count,
-                postcode: @form.model.pretty_postcode) %>
-        <% end %>
-
-        <%= govuk_link_to(
-              t(".search_again"),
-              providers_legal_aid_application_address_lookup_path(@legal_aid_application),
-              class: "govuk-body change-link change-postcode-link",
-              "aria-label": "postcode",
-            ) %>
-      </p>
-
-      <div class="govuk-inset-text">
-        <%= sanitize @address_collection.first.address.gsub(",", "<br>") %>
-      </div>
-
-      <%= form.hidden_field :address_line_one, value: @addresses.first.address_line_one %>
-      <%= form.hidden_field :address_line_two, value: @addresses.first.address_line_two %>
-      <%= form.hidden_field :city, value: @addresses.first.city %>
-      <%= form.hidden_field :county, value: @addresses.first.county %>
-      <%= form.hidden_field :postcode, value: @addresses.first.postcode %>
-      <%= form.hidden_field :lookup_postcode, value: @addresses.first.postcode %>
-
-      <p><%= govuk_link_to t(".link_text"), providers_legal_aid_application_address_path(@legal_aid_application) %></p>
-      <%= next_action_buttons(
-            continue_id: "select-address-button",
-            show_draft: false,
-            form:,
-            continue_button_text: t(".continue_button_text"),
-          ) %>
-<% end %>
