--- app/views/shared/home_address_selection/_no_address.html.erb
@@ -1,27 +1,0 @@
-<%= page_template(
-      page_title: t(".heading"),
-      template: :basic,
-      form:,
-    ) do %>
-
-      <h1 class="govuk-heading-xl"><%= page_title %></h1>
-
-      <p class="govuk-body">
-        <% if @form.model.building_number_name.present? %>
-          <%= t(".addresses_found_long_html",
-                number: @address_collection.count,
-                postcode: @form.model.pretty_postcode,
-                building_name_number: @form.model.building_number_name) %>
-        <% else %>
-          <%= t(".addresses_found_html",
-                number: @address_collection.count,
-                postcode: @form.model.pretty_postcode) %>
-        <% end %>
-      </p>
-
-      <p class="govuk-body"><%= t(".paragraph_2") %></p>
-
-      <p><%= govuk_link_to(t(".change_postcode"), providers_legal_aid_application_address_lookup_path) %></p>
-      <p><%= govuk_link_to(t(".manual_address"), providers_legal_aid_application_address_path) %></p>
-      <p><%= govuk_button_link_to(t("generic.save_and_come_back_later"), providers_legal_aid_applications_path, secondary: true) %></p>
-<% end %>
