+++ config/routes.rb
@@ -176,6 +176,8 @@ Rails.application.routes.draw do
       namespace :home_address do
         resource :different_address, only: %i[show update], path: "correspondence_is_home_address"
         resource :different_address_reason, only: %i[show update], path: "why_addresses_differ"
+        resource :home_address_lookup, only: %i[show update], path: "find_home_address"
+        resource :home_address_selection, only: %i[show update], path: "address_results"
       end
       resource :delete, controller: :delete, only: %i[show destroy]
       resources :proceedings_types, only: %i[index create]
@@ -184,9 +186,7 @@ Rails.application.routes.draw do
       resource :applicant_details, only: %i[show update]
       resource :address, only: %i[show update], path: "enter_correspondence_address"
       resource :address_lookup, only: %i[show update], path: "find_correspondence_address"
-      resource :home_address_lookup, only: %i[show update], path: "find_home_address"
       resource :address_selection, only: %i[show update], path: "correspondence_address_results"
-      resource :home_address_selection, only: %i[show update], path: "home_address_results"
       resource :check_benefit, only: %i[index update]
       resource :has_national_insurance_number, only: %i[show update]
       resources :check_benefits, only: [:index]
