+++ spec/requests/providers/home_address/home_address_lookups_controller_spec.rb
@@ -1,12 +1,12 @@
 require "rails_helper"
 
-RSpec.describe Providers::HomeAddressLookupsController do
+RSpec.describe Providers::HomeAddress::HomeAddressLookupsController do
   let(:legal_aid_application) { create(:legal_aid_application, :with_applicant) }
   let(:applicant) { legal_aid_application.applicant }
   let(:provider) { legal_aid_application.provider }
 
-  describe "GET /providers/applications/:legal_aid_application_id/home_address_lookup" do
-    subject(:get_request) { get providers_legal_aid_application_home_address_lookup_path(legal_aid_application) }
+  describe "GET /providers/applications/:legal_aid_application_id/home_address/home_address_lookup" do
+    subject(:get_request) { get providers_legal_aid_application_home_address_home_address_lookup_path(legal_aid_application) }
 
     context "when the provider is not authenticated" do
       before { get_request }
@@ -22,13 +22,13 @@ RSpec.describe Providers::HomeAddressLookupsController do
 
       it "shows the postcode entry page" do
         expect(response).to be_successful
-        expect(unescaped_response_body).to include(I18n.t("providers.home_address_lookups.show.heading"))
+        expect(unescaped_response_body).to include("Find your client's home address")
       end
     end
   end
 
-  describe "PATCH/providers/applications/:legal_aid_application_id/home_address_lookup" do
-    subject(:patch_request) { patch providers_legal_aid_application_home_address_lookup_path(legal_aid_application), params: }
+  describe "PATCH/providers/applications/:legal_aid_application_id/home_address/home_address_lookup" do
+    subject(:patch_request) { patch providers_legal_aid_application_home_address_home_address_lookup_path(legal_aid_application), params: }
 
     let(:postcode) { "SW1H 9EA" }
     let(:normalized_postcode) { "SW1H9AE" }
@@ -78,7 +78,7 @@ RSpec.describe Providers::HomeAddressLookupsController do
 
         it "redirects to the home address selection page" do
           patch_request
-          expect(response).to redirect_to(providers_legal_aid_application_home_address_selection_path)
+          expect(response).to redirect_to(providers_legal_aid_application_home_address_home_address_selection_path)
         end
 
         context "and a building number" do
